Add build flags from BuildResultHelper to TestUiSession
With cl/478816486, the responsibility of adding a build flag corresponding to the path of the local bep file was moved from testUiSessionProvider to the buildResultHelper. As a result, while creating a testUiSession, we need to add buildResultHelper.getBuildFlags() to the blaze flags.

The test debugging runConfig uses BlazeJavaRunProfileState and buildResultHelper.getBuildFlags() is missing in it causing the tests to fail when debugged. This CL fixes that.

PiperOrigin-RevId: 483457631